# Faster unmarshal (30x faster) in some case