--- a/benchmarks/benchmark_test.go
+++ b/benchmarks/benchmark_test.go
@@ -1,6 +1,8 @@
 package benchmarks
 
 import (
+	"encoding/hex"
+	"strings"
 	"testing"
 
 	"github.com/goccy/go-yaml"
@@ -47,6 +49,77 @@ elements:
 			if err := yaml.Unmarshal([]byte(src), &t); err != nil {
 				b.Fatal(err)
 			}
+		}
+	})
+}
+
+type Address [20]byte
+
+func (a *Address) UnmarshalText(text []byte) error {
+	decoded, err := hex.DecodeString(strings.TrimPrefix(string(text), "0x"))
+	if err != nil {
+		return err
+	}
+	copy(a[:], decoded)
+	return nil
+}
+
+func BenchmarkUnmarshalBigDocument(b *testing.B) {
+	type Token struct {
+		Address Address `yaml:"address"`
+		Name    string  `yaml:"name"`
+	}
+
+	type D struct {
+		Tokens []Token `yaml:"tokens"`
+	}
+
+	expAddr := Address{0x12, 0x34, 0x56, 0x78, 0x90, 0xab, 0xcd, 0xef, 0x12, 0x34, 0x56, 0x78, 0x90, 0xab, 0xcd, 0xef, 0x12, 0x34, 0x56, 0x78}
+
+	getDoc := func(n int) string {
+		doc := "tokens:\n"
+		for i := 0; i < n; i++ {
+			doc += "  - address: 0x1234567890abcdef1234567890abcdef12345678\n"
+			doc += "    name: token\n"
+		}
+		return doc
+	}
+
+	src := getDoc(2_000)
+
+	b.Run("gopkg.in/yaml.v2", func(b *testing.B) {
+		var t D
+		for i := 0; i < b.N; i++ {
+			if err := goyaml2.Unmarshal([]byte(src), &t); err != nil {
+				b.Fatal(err)
+			}
+		}
+		if t.Tokens[0].Address != expAddr {
+			b.Fatal("invalid address")
+		}
+	})
+
+	b.Run("gopkg.in/yaml.v3", func(b *testing.B) {
+		var t D
+		for i := 0; i < b.N; i++ {
+			if err := goyaml3.Unmarshal([]byte(src), &t); err != nil {
+				b.Fatal(err)
+			}
+		}
+		if t.Tokens[0].Address != expAddr {
+			b.Fatal("invalid address")
+		}
+	})
+
+	b.Run("github.com/goccy/go-yaml", func(b *testing.B) {
+		var t D
+		for i := 0; i < b.N; i++ {
+			if err := yaml.Unmarshal([]byte(src), &t); err != nil {
+				b.Fatal(err)
+			}
+		}
+		if t.Tokens[0].Address != expAddr {
+			b.Fatal("invalid address")
 		}
 	})
 }