--- a/benchmarks/go.mod
+++ b/benchmarks/go.mod
@@ -1,6 +1,8 @@
 module benchmarks
 
-go 1.12
+go 1.21.0
+
+toolchain go1.23.8
 
 replace github.com/goccy/go-yaml => ../