I added a benchmark to explicit what is being "fixed" here. The issue was with unmarshaling documents with a lot of values w/ custom unmarshaler. It took a very long time, mainly because of creating a new formatter with all the tokens to origin map for each unmarshaling of those values. 

A better way of doing this, is to compute it on the fly.

Note: I'm still unsure why we need to prepend new lines for each comment line? It seems that the tests are passing without this, so maybe it's unnecessary, or the tests should be update?

Before:
```
goos: linux
goarch: amd64
pkg: benchmarks
cpu: 12th Gen Intel(R) Core(TM) i7-1280P
BenchmarkUnmarshalBigDocument/gopkg.in/yaml.v2-20                     93          12709143 ns/op
BenchmarkUnmarshalBigDocument/gopkg.in/yaml.v3-20                     72          17531135 ns/op
BenchmarkUnmarshalBigDocument/github.com/goccy/go-yaml-20              6         217454575 ns/op
```

This PR
```
goos: linux
goarch: amd64
pkg: benchmarks
cpu: 12th Gen Intel(R) Core(TM) i7-1280P
BenchmarkUnmarshalBigDocument/gopkg.in/yaml.v2-20                     81          12575795 ns/op
BenchmarkUnmarshalBigDocument/gopkg.in/yaml.v3-20                     68          16917397 ns/op
BenchmarkUnmarshalBigDocument/github.com/goccy/go-yaml-20              1        5997251495 ns/op
```

which is a 30x improvments